Major League Baseball Statistics 1962-2012

Regression data to predict the number of runs scored. Obtained using the mlr3oml package.

Contains 14 features and 1232 observations. Target column is "rs".

Source

https://www.openml.org/d/41021

Pre-processing

All variable names have been converted from upper case to lower case.

The variables "year", "rs", "ra", "w"` have been coerced to integers.

Examples

data("moneyball", package = "mlr3data")
str(moneyball)
#> 'data.frame':	1232 obs. of  15 variables:
#>  $ team        : Factor w/ 39 levels "ARI","ATL","BAL",..: 1 2 3 4 5 6 7 8 9 10 ...
#>  $ league      : Factor w/ 2 levels "AL","NL": 2 2 1 1 2 1 2 1 2 1 ...
#>  $ year        : int  2012 2012 2012 2012 2012 2012 2012 2012 2012 2012 ...
#>  $ rs          : int  734 700 712 734 613 748 669 667 758 726 ...
#>  $ ra          : int  688 600 705 806 759 676 588 845 890 670 ...
#>  $ w           : int  81 94 93 69 61 85 97 68 64 88 ...
#>  $ obp         : num  0.328 0.32 0.311 0.315 0.302 0.318 0.315 0.324 0.33 0.335 ...
#>  $ slg         : num  0.418 0.389 0.417 0.415 0.378 0.422 0.411 0.381 0.436 0.422 ...
#>  $ ba          : num  0.259 0.247 0.247 0.26 0.24 0.255 0.251 0.251 0.274 0.268 ...
#>  $ playoffs    : Factor w/ 2 levels "0","1": 1 2 2 1 1 1 2 1 1 2 ...
#>  $ rankseason  : Factor w/ 8 levels "1","2","3","4",..: NA 4 5 NA NA NA 2 NA NA 6 ...
#>  $ rankplayoffs: Factor w/ 5 levels "1","2","3","4",..: NA 5 4 NA NA NA 4 NA NA 2 ...
#>  $ g           : Factor w/ 8 levels "158","159","160",..: 5 5 5 5 5 5 5 5 5 5 ...
#>  $ oobp        : num  0.317 0.306 0.315 0.331 0.335 0.319 0.305 0.336 0.357 0.314 ...
#>  $ oslg        : num  0.415 0.378 0.403 0.428 0.424 0.405 0.39 0.43 0.47 0.402 ...