Palmer Penguins

Classification data to predict the species of penguins from the palmerpenguins package (CRAN version 0.1.0). Promoted as an alternative to the iris data set.

Source

palmerpenguins

Pre-processing

The unit have been removed from the column names. Lengths are given in millimeters (mm), weight in gram (g).

References

Gorman KB, Williams TD, Fraser WR (2014). “Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis).” PLoS ONE, 9(3), e90081. doi: 10.1371/journal.pone.0090081 .

https://github.com/allisonhorst/palmerpenguins

Examples

data("penguins", package = "mlr3data")
str(penguins)

#> 'data.frame':	344 obs. of  8 variables:
#>  $ species       : Factor w/ 3 levels "Adelie","Chinstrap",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ island        : Factor w/ 3 levels "Biscoe","Dream",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ bill_length   : num  39.1 39.5 40.3 NA 36.7 39.3 38.9 39.2 34.1 42 ...
#>  $ bill_depth    : num  18.7 17.4 18 NA 19.3 20.6 17.8 19.6 18.1 20.2 ...
#>  $ flipper_length: int  181 186 195 NA 193 190 181 195 193 190 ...
#>  $ body_mass     : int  3750 3800 3250 NA 3450 3650 3625 4675 3475 4250 ...
#>  $ sex           : Factor w/ 2 levels "female","male": 2 1 1 NA 1 2 1 2 NA NA ...
#>  $ year          : int  2007 2007 2007 2007 2007 2007 2007 2007 2007 2007 ...